added liveUpdates
