+++ src/lib/components/ShexFormBuilder/children/ShexFormLive/index.js
@@ -1,0 +1,3 @@
+import ShexFormLive from './shex-form-live.component';
+
+export { ShexFormLive };
